std::clamp -> amrex::Clamp (#4292)
std::clamp is a constexpr function, and it is supposed to work in device
code. However, std::clamp in gcc 14 seems to be incompatible with clang.
This is causing clang based HIP compiler to fail on systems using gcc 14
libstd as reported by multiple users. Here we work around the issue by
using amrex::Clamp instead.