Use bigint in HypreSolver test (#4297)
Also fix a bug when HYPRE_Int is long long. The issue was in the
previous code the discard branch of `if constexpr` was still checked by
the compiler, because it's not in a function template, even though it's
in a member function of a class template. We now turn the function into
a template.